# modulo_2_SI/2.1/prac2_bdd/Java/d.reyg-swagger.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/users/": {
      "get": {
        "tags": [
          "user-rest-controller"
        ],
        "summary": "Get all users",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Found all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user-rest-controller"
        ],
        "summary": "Update information about user",
        "operationId": "replaceUser",
        "requestBody": {
          "description": "User to be updated",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user attributes supplied"
          }
        }
      },
      "post": {
        "tags": [
          "user-rest-controller"
        ],
        "summary": "Create a new user",
        "operationId": "createUser",
        "requestBody": {
          "description": "User to be created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user attributes supplied"
          },
          "406": {
            "description": "Nick user exists"
          }
        }
      }
    },
...
    "/api/v1/users/{userId}": {
      "get": {
        "tags": [
          "user-rest-controller"
        ],
        "summary": "Get a user by its id",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "id of user to be searched",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found the user and her comments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid format id supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "user-rest-controller"
        ],
        "summary": "Delete a user",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "identifier of the user",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted user without comment associated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid format id supplied"
          },
          "404": {
            "description": "User not found"
          },
          "406": {
            "description": "User with comments associates"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserRequestDto": {
        "required": [
          "nick"
        ],
        "type": "object",
        "properties": {
          "nick": {
            "type": "string"
          },
          "mail": {
            "type": "string"
          }
        }
      },
      "CommentResponseUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "comment": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "nick": {
            "type": "string"
          },
          "mail": {
            "type": "string"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponseUserDto"
            }
          }
        }
      },
...
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "comment": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
    }
  }
}